+++ config/services/core.mjs
@@ -29,7 +29,7 @@ export const resolveServiceConfiguration = ({ utils }) => {
       // Image to run (different in dev)
       image: PROD ? 'itsmorio/core' : utils.isUnitTest() ? 'itsmorio/core-test' : 'itsmorio/core-dev',
       // Image tag (version) to run
-      tag: utils.getPreset('MORIO_VERSION'),
+      tag: utils.getPreset('MORIO_VERSION_TAG'),
       // Don't attach to the default network
       networks: { default: null },
       // Ports to export (not in production)