+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Prefix the version number with v in git and container image tags
+
 ## [0.4.0] - 2024-09-25
 
 ### Added